Remove HOCs and proptypes
